+++ wp-content/plugins/rollback-update-failure/plugin.php
@@ -3,20 +3,19 @@
  * Rollback Update Failure
  *
  * @package rollback-update-failure
- * @author Andy Fragen <andy@thefragens.com>, Ari Stathopolous <aristath@gmail.com>
  * @license MIT
  */
 
 /**
  * Plugin Name: Rollback Update Failure
- * Author: Andy Fragen, Ari Stathopolous, Colin Stewart, Paul Biron
+ * Author: WP Core Contributors
  * Description: Feature plugin to test plugin/theme update failures and rollback to previous installed packages.
- * Version: 3.3.1
+ * Version: 5.0.2
  * Network: true
  * License: MIT
  * Text Domain: rollback-update-failure
  * Requires PHP: 5.6
- * Requires at least: 6.0
+ * Requires at least: 6.2
  * GitHub Plugin URI: https://github.com/WordPress/rollback-update-failure
  * Primary Branch: main
  */
@@ -25,24 +24,17 @@ namespace Rollback_Update_Failure;
 
 /*
  * Exit if called directly.
- * PHP version check and exit.
  */
 if ( ! defined( 'WPINC' ) ) {
 	die;
 }
 
-// Deactivate plugin when committed to core.
-if ( version_compare( get_bloginfo( 'version' ), '6.2-beta1', '>=' ) ) {
+// Hooray move_dir() has been committed.
+if ( version_compare( get_bloginfo( 'version' ), '6.2-beta1', '<=' ) ) {
 	require_once ABSPATH . 'wp-admin/includes/plugin.php';
 	deactivate_plugins( __FILE__ );
+	return;
 }
 
-// Load files.
-require_once __DIR__ . '/wp-admin/includes/class-wp-site-health.php';
-require_once __DIR__ . '/wp-admin/includes/class-plugin-theme-upgrader.php';
-require_once __DIR__ . '/wp-admin/includes/class-wp-upgrader.php';
-require_once __DIR__ . '/wp-admin/includes/file.php';
-require_once __DIR__ . '/wp-includes/update.php';
-
-// For testing.
-require_once __DIR__ . '/testing/failure-simulator.php';
+// Load the Composer autoloader.
+require __DIR__ . '/vendor/autoload.php';